Reorganize utils.
