+++ lib/compile/bundles/bootstrap.js
@@ -4,7 +4,7 @@ import * as Pluggable from "../../pluggable";
 import resolveAsset from "../modules/resolve";
 import loadAst from "../modules/load-ast";
 
-import { entries } from "../../util";
+import { entries } from "../../util/object";
 
 const bootstrapBundle = Pluggable.promise(function bootstrapBundle (srcPath, bundlDef, isEntryPt) {
   return this.resolveAsset(srcPath)
