+++ lib/compile/modules/compile.js
@@ -5,7 +5,7 @@ import _ from "lodash";
 import { transform } from "babel-core";
 
 import * as Pluggable from "../../pluggable";
-import { stream as streamUtils } from "../../util";
+import { toArray, toHash } from "../../util/stream";
 
 import resolve from "./resolve";
 import loadAst from "./load-ast";
@@ -44,9 +44,9 @@ const compileModules = Pluggable.stream(function compileModules (seedModules) {
     const descendants = directDeps.flatMap(dep => genModules(dep));
 
     const hashedModulePromise = Promise.all([
-      streamUtils.toArray(directDeps),
-      streamUtils.toArray(descendants),
-      streamUtils.toHash(descendants, "path")
+      toArray(directDeps),
+      toArray(descendants),
+      toHash(descendants, "path")
     ])
       .then(([dependencies, deepDependencies, deepDepsHash]) => {
         dependencies = dependencies.map(dep => deepDepsHash[dep.path]);
