+++ lib/index.js
@@ -7,7 +7,7 @@ import most from "most";
 import _ from "lodash";
 
 import compile from "./compile";
-import { entries } from "./util";
+import { entries } from "./util/object";
 import loadAst from "./compile/modules/load-ast";
 import compileModules from "./compile/modules/compile";
