+++ lib/util/object.js
@@ -1,0 +1,5 @@
+export function* entries (obj) {
+  for (let key of Object.keys(obj)) {
+    yield [key, obj[key]];
+  }
+}
